Feat: Enable dynamic addition of menus in DropdownMenu component
Enhanced the DropdownMenu component to allow for the dynamic addition of menus, providing flexibility to include various menu options beyond Share and Report.
