+++ src/components/DropdownMenu.style.ts
@@ -21,16 +21,3 @@ export const dropdownMenuButtonStyle: SxProps = {
   boxSizing: 'border-box',
   p: '0.0625rem',
 }
-
-export const menuItemStyle: SxProps = {
-  p: 0,
-  height: '2.5rem',
-  width: '7.5rem',
-  boxSizing: 'border-box',
-}
-
-export const menuItemIconStyleBase: SxProps = {
-  width: '1.5rem',
-  height: '1.5rem',
-  boxSizing: 'border-box',
-}
