+++ src/components/DropdownMenu.tsx
@@ -1,61 +1,12 @@
 'use client'
 import React from 'react'
-import { Button, Menu, MenuItem, Stack, Typography, Fade } from '@mui/material'
-import ShareIcon from '@/icons/ShareIcon'
+import { Button, Menu, MenuItem, Fade } from '@mui/material'
 import MoreHorizontalIcon from '@/icons/MoreHorizontalIcon'
-import ReportIcon from '@/icons/ReportIcon'
 import useMedia from '@/hook/useMedia'
 import * as style from './DropdownMenu.style'
+import * as dropdownMenuStyle from './dropdownMenu/dropdownMenu.styles'
 
-const IconMenuItem = ({
-  icon,
-  text,
-  action,
-}: {
-  icon: React.ReactNode
-  text: string
-  action?: () => void
-}) => {
-  const handleClick = () => {
-    if (action) {
-      action()
-    }
-  }
-  return (
-    <MenuItem dense onClick={handleClick}>
-      <Stack
-        direction={'row'}
-        spacing={'0.375rem'}
-        alignItems={'center'}
-        justifyContent={'center'}
-        sx={style.menuItemStyle}
-      >
-        {icon}
-        <Typography variant="Caption" color={'text.alternative'}>
-          {text}
-        </Typography>
-      </Stack>
-    </MenuItem>
-  )
-}
-
-// TODO : 신고하기 기능 구현, 수정하기 버튼 구현
-// TODO : toast 동작하도록 처리
-const DropdownMenu = ({
-  title,
-  url,
-  content,
-}: {
-  title: string
-  url: string
-  content: string
-  // COMMENT : #460이 머지되어야 사용 가능
-  message?: string // 클립보드에 복사할 메세지 ex) '피어에서 동료를 구해보세요! 이런 프로젝트가 있어요! ${url}'
-  setInfoToastMessage?: (message: string) => void
-  openInfoToast?: () => void
-  setErrorToastMessage?: (message: string) => void
-  openErrorToast?: () => void
-}) => {
+const DropdownMenu = ({ children }: { children: React.ReactNode }) => {
   const [anchorEl, setAnchorEl] = React.useState<null | HTMLElement>(null)
   const open = Boolean(anchorEl)
   const { isPc } = useMedia()
@@ -67,52 +18,6 @@ const DropdownMenu = ({
 
   const handleClose = () => {
     setAnchorEl(null)
-  }
-
-  const handleShare = () => {
-    if (navigator.share) {
-      navigator
-        .share({
-          title: title,
-          url: url,
-          text: content,
-        })
-        .then(() => {
-          handleClose()
-          // setInfoToastMessage('공유하기가 완료되었습니다.')
-          // openInfoToast()
-        })
-        .catch((e) => {
-          console.error(e)
-          // setErrorToastMessage('공유를 실패했습니다.')
-          // openErrorToast()
-        })
-    } else {
-      // COMMENT : #460이 머지된 이후 message를 사용하도록 수정
-      const copyText = `피어에서 동료를 구해보세요! 이런 프로젝트가 있어요:)\n[${title}]\n${url}\n`
-
-      if (navigator.clipboard) {
-        navigator.clipboard
-          .writeText(copyText)
-          .then(() => alert('클립보드에 복사되었습니다.'))
-          .catch((e) => {
-            console.error(e)
-            // setErrorToastMessage('클립보드에 복사하지 못했습니다.')
-            // openErrorToast()
-          })
-      } else {
-        const textarea = document.createElement('textarea')
-        textarea.value = copyText
-        textarea.id = 'temp'
-        document.body.appendChild(textarea)
-        textarea.select()
-        document.execCommand('copy')
-        document.body.removeChild(textarea)
-        alert('클립보드에 복사되었습니다.')
-      }
-
-      handleClose()
-    }
   }
 
   return (
@@ -156,7 +61,7 @@ const DropdownMenu = ({
       >
         <MenuItem
           onClick={handleClose}
-          sx={{ ...style.menuItemStyle, position: 'relative' }}
+          sx={{ ...dropdownMenuStyle.menuItemStyle, position: 'relative' }}
         >
           <MoreHorizontalIcon
             sx={{
@@ -168,28 +73,7 @@ const DropdownMenu = ({
             }}
           />
         </MenuItem>
-        <IconMenuItem
-          action={handleShare}
-          icon={
-            <ShareIcon
-              sx={{
-                ...style.menuItemIconStyleBase,
-                padding: '0.125rem',
-                color: 'text.alternative',
-              }}
-            />
-          }
-          text={'공유'}
-        />
-        <IconMenuItem
-          action={handleClose}
-          icon={
-            <ReportIcon
-              sx={{ ...style.menuItemIconStyleBase, color: 'text.alternative' }}
-            />
-          }
-          text={'신고'}
-        />
+        {children}
       </Menu>
     </div>
   )
