+++ src/components/dropdownMenu/dropdownMenu.styles.ts
@@ -1,0 +1,14 @@
+import { SxProps } from '@mui/material'
+
+export const menuItemStyle: SxProps = {
+  p: 0,
+  height: '2.5rem',
+  width: '7.5rem',
+  boxSizing: 'border-box',
+}
+
+export const menuItemIconStyleBase: SxProps = {
+  width: '1.5rem',
+  height: '1.5rem',
+  boxSizing: 'border-box',
+}
